+++ application.js
@@ -26,7 +26,7 @@ app.use(mbaasExpress.fhmiddleware());
 
 // fhlint-begin: custom-routes
 app.use('/try', require('./lib/try.js')());
-app.use('/test', require('./lib/test.js')());
+app.use('/test', require('./lib/tester.js')());
 // fhlint-end
 
 // Important that this is last!
