+++ lib/tester.js
