+++ public/index.html
@@ -1,12 +1,12 @@
 <html>
   <head>
     <title>API Mapper</title>
-    <link rel="stylesheet" href="css/main.css" />
+    <link rel="stylesheet" href="css/postie.css" />
   </head>
   <body>
     <form class="try">
 
-      <h2>Request1</h2>
+      <h2>Request</h2>
 
       <div class="row">
         <select name="method">
@@ -58,6 +58,6 @@
 
 
     <script src="http://code.jquery.com/jquery-2.1.4.min.js"></script>
-    <script src="js/main.js"></script>
+    <script src="js/postie.js"></script>
   </body>
 </html>
